Refactor get_or_fetch methods to utils file
